[레거시 코드 리팩터링 - 3단계] 루키(최지훈) 미션 제출합니다. (#396)
* refactor: 메뉴 상품의 실제 상품 가격에 대한 값에 @Transient 적용

* feat: 메뉴의 정보가 변경되어도 주문메뉴의 정보가 변경되지 않도록 추가

* refactor: BigDecimal로 가격의 요청타입을 변경

* refactor: 도메인의 패키지 구조 변경

* refactor: 데이터 클리너 설정 변경

* feat: Event 객체 설정

* refactor: Event를 통해 주문 생성, 상태 변경시 주문 테이블의 상태를 변경하도록 구현

* refactor: 각 도메인 패키지에 Repository를 이동

* refactor: Table에서 Order를 바라보는 의존 제거

* refactor: Core 패키지를 Common 패키지로 변경

* refactor: 사용하지 않는 import 제거

* refactor: Order 생성을 하는 정적 팩터리 메서드 추가

* feat: 주문 테이블에 여러개의 주문이 있을 때에 대한 검증 로직 추가

* fix: 접근제어자 private 으로 변경

+++ src/main/java/kitchenpos/application/ProductService.java
@@ -1,11 +1,10 @@
 package kitchenpos.application;
 
-import java.math.BigDecimal;
 import java.util.stream.Collectors;
 import kitchenpos.application.dto.ProductCreateRequest;
 import kitchenpos.application.dto.ProductResponse;
-import kitchenpos.domain.Product;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.domain.product.Product;
+import kitchenpos.domain.product.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,7 +20,7 @@ public class ProductService {
     }
 
     public ProductResponse create(final ProductCreateRequest request) {
-        final Product product = productRepository.save(new Product(request.getName(), BigDecimal.valueOf(request.getPrice())));
+        final Product product = productRepository.save(new Product(request.getName(), request.getPrice()));
 
         return ProductResponse.createResponse(product);
     }
